+++ src/main/java/nl/bakkerv/p1/parser/text/WattValueParser.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class WattValueParser implements ValueParser<Integer> {
+public class WattValueParser implements ValueParser<BigDecimal> {
 
 	private Pattern pattern;
 
@@ -14,7 +14,7 @@ public class WattValueParser implements ValueParser<Integer> {
 	}
 
 	@Override
-	public Optional<TimestampedValue<Integer>> parse(final String value) {
+	public Optional<TimestampedValue<BigDecimal>> parse(final String value) {
 		BigDecimal result = null;
 
 		Matcher matcher = this.pattern.matcher(value);
@@ -23,6 +23,6 @@ public class WattValueParser implements ValueParser<Integer> {
 			result = new BigDecimal(matcher.group(1)).multiply(new BigDecimal(1000)).setScale(0);
 		}
 
-		return result == null ? Optional.empty() : Optional.of(new TimestampedValue<>(Optional.empty(), result.intValue()));
+		return result == null ? Optional.empty() : Optional.of(new TimestampedValue<>(Optional.empty(), result));
 	}
 }
